v0.1.2 I-8: a surface radius is a world position, so hold it at float64
`Chunk.surface` kept each column's ground and water radius in a Float32Array.
At the shipped planet's 6,800 m a float32 step is 0.49 mm, so storing one
rounded it: the cells inside a chunk were read back rounded, and the cells
around a chunk are generated on demand at float64, which left one flat surface
with two readings 48 micrometres apart.

`layerOfSurface` is a ceil. Two readings that straddle a layer boundary name
different layers however small the gap, and the mesher walls the drop -- a
whole block of cliff along the rim of every chunk. The paused world puts every
column on a boundary exactly, because its ground is at sea level and sea level
is a radius, so the cliffs stood everywhere: 380 wall triangles a chunk,
shaded as sides and occluded at their corners, which is the dark line grid and
the dark band across the horizon.

Holding the radius at float64 costs 4,488 bytes a chunk against blocks of
478 KB. Rounding the generated columns to match instead would keep a world
position at float32, which invariant 13 forbids for this reason, and makes
every future reader remember to round the same way.

Two tests hold it: a stored radius is bit-identical to the generator's on a
planet float32 cannot represent, and a world whose surface tops a layer
exactly meshes to caps alone. Both fail on the float32 array. Over the bench's
paused scenes the cliffs were 13% of every triangle drawn.

precision.js gains section 10 and doc 15 the subsection it verifies; the plan
carries I-8 and the register F-027, which corrects F-011 -- this container can
present WebGPU after all, and reading the frame is what found this.

+++ packages/engine/src/generation/chunk/Chunk.ts
@@ -44,8 +44,16 @@ export class Chunk {
 	/**
 	 * Each slot's true surface radii, ground then water, before the rounding
 	 * to this chunk's own layer grid. Zeros where nothing recorded them.
+	 *
+	 * A radius is a world position, so it is held at `float64` however small
+	 * the number looks. The cells around a chunk are generated on demand and
+	 * arrive at full precision, and the rounding that turns a radius into a
+	 * layer is a `ceil`: two readings of one surface that differ by a
+	 * `float32` step at planet radius -- 0.49 mm at 6,800 m -- land on
+	 * different layers whenever the surface sits on a layer boundary, and a
+	 * whole block of cliff stands between them.
 	 */
-	readonly surface: Float32Array;
+	readonly surface: Float64Array;
 
 	constructor(
 		address: ChunkAddress,
@@ -54,7 +62,7 @@ export class Chunk {
 		layerCount: number,
 		blocks?: Uint16Array,
 		band?: Int16Array,
-		surface?: Float32Array,
+		surface?: Float64Array,
 	) {
 		this.address = address;
 		this.depth = depth;
@@ -64,7 +72,7 @@ export class Chunk {
 		this.layerCount = layerCount;
 		this.blocks = blocks ?? new Uint16Array(this.slots * layerCount);
 		this.band = band ?? new Int16Array(this.slots * 2);
-		this.surface = surface ?? new Float32Array(this.slots * 2);
+		this.surface = surface ?? new Float64Array(this.slots * 2);
 	}
 
 	/** The column at a slot, as the mesher reads it. */
